+++ tiberiandawn/dllinterface.cpp
@@ -6313,7 +6313,7 @@ void DLLExportClass::Calculate_Start_Positions(void)
     for (int i = 0; i < MPlayerCount; i++) {
         PlayerPtr = HouseClass::As_Pointer(MPlayerHouses[i]);
         if (PlayerPtr) {
-            long x, y;
+            int x, y;
             Map.Compute_Start_Pos(x, y);
             MultiplayerStartPositions[i] = XY_Cell(x, y);
         }
