+++ tiberiandawn/unit.h
@@ -162,7 +162,7 @@ public:
     **	AI.
     */
     virtual DirType Desired_Load_Dir(ObjectClass* passenger, CELL& moveto) const;
-    virtual RadioMessageType Receive_Message(RadioClass* from, RadioMessageType message, long& param);
+    virtual RadioMessageType Receive_Message(RadioClass* from, RadioMessageType message, int& param);
     virtual void AI(void);
     virtual int Mission_Attack(void);
     virtual int Mission_Unload(void);
